Дипломный проект - Этап 2 - 2 итерация

+++ ewm-service/src/main/java/ru/practicum/repositories/CompilationRepository.java
@@ -2,10 +2,10 @@ package ru.practicum.repositories;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.models.compilations.Compilation;
+import ru.practicum.models.compilation.Compilation;
 
 import java.util.List;
 
-public interface CompilationRepository extends JpaRepository<Compilation, Integer> {
+public interface CompilationRepository extends JpaRepository<Compilation, Long> {
     List<Compilation> findAllByPinned(Boolean pinned, PageRequest pageable);
 }
